% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_two_arm.R
\name{sim_trial_dat}
\alias{sim_trial_dat}
\title{Simulate trial data using Poisson process for accr}
\usage{
sim_trial_dat(
  p1tru = 0.1,
  p2tru = 0.1,
  nmax = 100,
  enro_rate = 1,
  enro_intensity = function(t) 1,
  resp_delay = function(n) stats::runif(n),
  simple_rand = TRUE
)
}
\arguments{
\item{p1tru}{True response rate under control}

\item{p2tru}{True response rate under treatment}

\item{nmax}{The maximum total sample size}

\item{enro_rate}{The baseline accrual rate}

\item{enro_intensity}{Function for changing accrual intensity}

\item{resp_delay}{Function returning response time value}

\item{simple_rand}{Use complete randomisation or assume equal allocation at each interim}
}
\value{
A data.table
}
\description{
Simulate trial data using Poisson process for accr
}
